requirements.txt
altgraph==0.17.4
DateTime==5.2
importlib-metadata==6.8.0
lxml==4.9.3
macholib==1.16.3
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3.post1
six==1.16.0
tk==0.1.0
TKinterModernThemes==1.10.0
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
zope.interface==6.1
